category service + menu
